Update config and attack logic for new experiment setup
Increase client count and adjust training parameters to optimize for scalability. Modify attack settings, including switching to untargeted attacks and disabling data transformations like cropping and Gaussian noise. Fix type mismatch in poisoning fraction handling in the client module.

+++ src/modules/client.py
@@ -34,7 +34,7 @@ class FlowerClient(fl.client.NumPyClient):
         self.model = ModelFactory.create_model(config).to(self.device)
         self.epsilon = config.ldp.epsilon
 
-        fraction = int(config.poisoning.fraction)
+        fraction = float(config.poisoning.fraction)
         r = np.random.random(1)[0]
         if r < fraction:
             print(f"client {client_id} is poisoned------------------------------------")
